+++ src/optimize_CGM/optimize_yield_up_jsim.cpp
@@ -78,7 +78,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
     //local_nd = 0.1;
 
     for (int m = 0; m < MONTE_CARLO; m++){
-        for (int r = 0; r < global_rand.size(); r++){
+        for (size_t r = 0; r < global_rand.size(); r++){
             global_rand[r] = rand_global_yield(local_nd);
         }
         //cout << endl;
@@ -140,7 +140,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
 
         if( opt->success != 0 ){    // 開発者確認用
             lambda = (MULTI_NUM - opt->success) / MULTI_NUM;
-            for(int i = 0; i < element.size(); i++){
+            for(size_t i = 0; i < element.size(); i++){
 
                 element[i].value = opt->sum_value[i] / opt->success + lambda * (element[i].value - opt->sum_value_f[i] / bunbo);   // 新たなパラメータに置き換える
                 //element[i].value = opt->sum_value[i] / opt->success;
@@ -159,7 +159,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
                 //not_upd = 0;
                 opt->suc_max = opt->success;
 
-                for(int i = 0; i < element.size(); i++){
+                for(size_t i = 0; i < element.size(); i++){
                     //element[i].value = opt->sum_value[i] / opt->success;
                     //element[i].value = opt->sum_value[i] / opt->success + lambda * (element[i].value - opt->sum_value_f[i] / bunbo);   // 新たなパラメータに置き換える
                     //opt->best_value[i] =element[i].value;
@@ -168,7 +168,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
         }
 
         //cout << reduce(begin(yield_his), end(yield_his)) / yield_his.size() << endl;
-        if(reduce(begin(yield_his), end(yield_his)) / yield_his.size() >= suc_th){  //直近 5 サイクルの歩留まり(success) の平均が 60 を超えたら、標準偏差を +0.01
+        if(reduce(begin(yield_his), end(yield_his)) / (int)yield_his.size() >= suc_th){  //直近 5 サイクルの歩留まり(success) の平均が 60 を超えたら、標準偏差を +0.01
             local_nd += 0.01;
             Margin_low(element,  jud, data_cir, arg_arr, 2);
             not_upd = 0;
@@ -176,7 +176,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
 
             cri_bias_sum = calc_score(element, power);
             if( cri_bias_sum > opt->cri_bias_best ){   // マージンの評価値 cri_bias_sum が最大だったらそれを現時点の最良の回路として置き換える
-                    for(int j = 0; j < element.size(); j++){
+                    for(size_t j = 0; j < element.size(); j++){
                         opt->best_value[j] = element[j].value;    // best_value配列に現時点の最良の回路のパラメータを格納
                     }
                     opt->best_value[element.size()] = Margin_num;  //best_value配列の最後の要素の次の要素に何番目の回路かを格納
@@ -194,7 +194,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
         
         //全て初期化
         opt->success = 0;
-        for(int i = 0; i < element.size(); i++){
+        for(size_t i = 0; i < element.size(); i++){
             opt->sum_value[i] = 0;
             opt->sum_value_f[i] = 0;
         }
@@ -209,7 +209,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
     cri_bias_sum = calc_score(element, power);    // cri_bias_sum = その回路のマージンの評価 (クリティカルマージン + 2 * バイアスマージン)
     //cout << "cri_bias_sum : " << cri_bias_sum << endl; 
     if( cri_bias_sum > opt->cri_bias_best ){   // マージンの評価値 cri_bias_sum が最大だったらそれを現時点の最良の回路として置き換える
-        for(int j = 0; j < element.size(); j++){
+        for(size_t j = 0; j < element.size(); j++){
             opt->best_value[j] = element[j].value;    // best_value配列に現時点の最良の回路のパラメータを格納
         }
         opt->best_value[element.size()] = Margin_num;  //best_value配列の最後の要素の次の要素に何番目の回路かを格納
@@ -217,7 +217,7 @@ void optimize_yield_up_jsim(vector<ele_unit> &element, vector<string> &data_cir,
     }
 
     cout <<  " This is the " << GetOrdinalNumber(static_cast<int>(opt->best_value[element.size()])) << " value." << endl;
-    for(int i = 0; i < element.size(); i++){
+    for(size_t i = 0; i < element.size(); i++){
         element[i].value = opt->best_value[i];  //最終的な最良の回路のパラメータをelement配列に格納
     }
